% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bel2mel.f.R
\name{bel2mel.f}
\alias{bel2mel.f}
\title{Bimodal edgelist to monomodal edgelist}
\usage{
bel2mel.f(dbl2bel = NULL, subset = NULL, type = c("utel", "crel"),
  out = stop("Specify output directory for your project."),
  check.for.saved.output = F, write2disk = F)
}
\arguments{
\item{write2disk}{}
}
\description{
Recieves a dbl2bel and outputs one or both of two mode projections in edgelist format. May be redundant with igraph::bipartite.projection.
}
