% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comps2cos.f.R
\name{comps2cos.f}
\alias{comps2cos.f}
\title{Components to Cosparallel Fast Network Clique Percolation}
\usage{

  comps2cos.f(mel2comps.dir = stop("Specify directory where mel2comps.txt edgelists are located."),
  cosparallel.path = "cos", threads = parallel::detectCores() - 1,
  window.buffer = system("free -b | awk 'NR==2 {print $6}'", intern = T) \%>\%
  as.numeric \%>\% log2)
}
\arguments{
\item{cosparallel.path}{A path to cosparallel executable (e.g. ~/cosparallel-code/cos). Default assumes cos executable is in path.}

\item{threads}{}
}
\description{
Components to Cosparallel Fast Network Clique Percolation
}
\examples{
\dontrun{
   # available - free memory for linux
   window.buffer=system('free -b | awk \\'NR==2 {print $3" "$6}\\'',intern = T) \%>\% strsplit(' ') \%>\% unlist \%>\% as.numeric \%>\% diff \%>\% log2
}
}
