% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cos2kcliqdb.f.R
\name{cos2kcliqdb.f}
\alias{cos2kcliqdb.f}
\title{Title}
\usage{

  cos2kcliqdb.f(mel2comps.dir = stop("Specify a mel2comps directory that includes cos output."),
  out = stop("Specify output directory."), type = c("crel", "utel"))
}
\arguments{
\item{type}{}
}
\description{
Title
}
