% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbl2bel.f.R
\name{dbl2bel.f}
\alias{dbl2bel.f}
\title{Title}
\usage{
dbl2bel.f(wok2dbl, out = stop("Specify output directory for your project."),
  check.for.saved.output = F, saved_recode = NULL, save_og4recode = F,
  trim_doi = T, capitalize = T, trim_anon = T, cut_samp_def = 0,
  trim_pendants = T, trim_loops = T)
}
\arguments{
\item{trim_loops}{}
}
\description{
Title
}
