% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbl2w.f.R
\name{dbl2w.f}
\alias{dbl2w.f}
\title{Title}
\usage{
dbl2w.f(wok2dbl, out = stop("Specify output directory"),
  field = stop("field=c(\\"field1\\",\\"field2\\",...)"), variations = NULL,
  recode = NULL)
}
\arguments{
\item{recode}{}
}
\description{
Title
}
