% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftxt2stmbow.f.R
\name{ftxt2stmbow.f}
\alias{ftxt2stmbow.f}
\title{Converts full text data to stm style document term matrix. Redundant with stm function textProcessor()}
\usage{
ftxt2stmbow.f(source.dir = NULL, string = NULL, names = NULL, lt = 1,
  save.to.disk = F, check.for.saved.output = F, out.dir,
  sample.docs = NULL)
}
\arguments{
\item{source.dir}{A path to a folder containing a plain text file for each document.}

\item{string}{A character string where each element is a full document.}

\item{names}{A vector of document names. Coerced to character. If unspecified no names will be given. If unspecified and source given, files names will be assigned instead.}
}
\description{
Converts full text data to stm style document term matrix. Redundant with stm function textProcessor()
}
