% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbng2tts.f.R
\name{gbng2tts.f}
\alias{gbng2tts.f}
\title{google books ngram 2 token time series}
\usage{
gbng2tts.f(query = c("social", "cultural"), out = "~/mat/dp", ys = 1901,
  ye = 2000, cfso = T, slug = "auto", sm = 0)
}
\arguments{
\item{query}{A character vector of terms or a long format database where the last column is the term and leading columns are factors. Factors will be returned on the result.}

\item{out}{hot stuff}

\item{ye}{}
}
\description{
google books ngram 2 token time series
}
