% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jstor2dbw.f.R
\name{jstor2dbw.f}
\alias{jstor2dbw.f}
\title{Title}
\usage{

  jstor2dbw.f(dir = stop("Choose input directory containing DFR.JSTOR.ORG batches of zip archives."),
  out = stop("Specify output directory for your project."), save = T,
  import.ngrams = F, sample.batches = T, sample.size = 1,
  in.parallel = F, drop.nchar1 = T, drop.freq1 = T, drop.doc1 = T,
  check.for.saved.output = F)
}
\arguments{
\item{check.for.saved.output}{}
}
\description{
Title
}
