% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jstorm2tab.f.R
\name{jstorm2tab.f}
\alias{jstorm2tab.f}
\title{JSTOR Master list to Table output}
\usage{
jstorm2tab.f(jstorm, beg.bef = 1900, end.aft = 2000, lan = "ENGLISH",
  pat = "(anth)|(soci[oa])|([^a-z]poli)|(econ)", ...)
}
\arguments{
\item{...}{Additional arguments to knitr::kable.}
}
\description{
JSTOR Master list to Table output
}
