% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcliqdb2viz.f.R
\name{kcliqdb2viz.f}
\alias{kcliqdb2viz.f}
\title{Title}
\usage{
kcliqdb2viz.f(cos2kcliqdb,
  mel2comps.dir = stop("Specify a mel2comps directory that includes cos output."),
  out = stop("Specify output directory."), type = c("crel", "utel"))
}
\arguments{
\item{type}{}
}
\description{
Title
}
