% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda2ftxt.f.R
\name{lda2ftxt.f}
\alias{lda2ftxt.f}
\title{Color coding original documents in topic highlighting.}
\usage{
lda2ftxt.f(map, doc.top, top.word, lda2rel, intensify = T, intensity = 0.3,
  sample = 10, out.dir = NULL, index = 0, spacing = 0.4, ptsize = 10,
  axes = F, pdf = F, mxdoc.word.prop = 0, fname, top.ord)
}
\description{
Color coding original documents in topic highlighting.
}
