% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda2rel.f.R
\name{lda2rel.f}
\alias{lda2rel.f}
\title{Original authors code to compute relevance. 'Forked' from https://github.com/cpsievert/LDAvis/blob/6f93aa85499b705c9ae6c56e5985df637f9f5132/R/createJSON.R}
\usage{
lda2rel.f(stmbow2lda, R = 10, lambda.step = 0.5, reorder.topics = FALSE,
  save.to.disk = F, check.for.saved.output = F, out.dir, ...)
}
\description{
Original authors code to compute relevance. 'Forked' from https://github.com/cpsievert/LDAvis/blob/6f93aa85499b705c9ae6c56e5985df637f9f5132/R/createJSON.R
}
