% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda2viz.f.R
\name{lda2viz.f}
\alias{lda2viz.f}
\title{Different approaches to reduce and visualize the standard topic-document and topic-word word matrices output by LDA and other topic modelling estimations.}
\usage{
lda2viz.f(stmbow2lda, out.dir, rt = F, ob = F, launch = T)
}
\description{
Different approaches to reduce and visualize the standard topic-document and topic-word word matrices output by LDA and other topic modelling estimations.
}
