% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master2jstorm.f.R
\name{master2jstorm.f}
\alias{master2jstorm.f}
\title{JSTOR Journal Master List}
\usage{

  master2jstorm.f(downloadurl = "http://www.jstor.org/kbart/collections/all-archive-titles?contentType=journals")
}
\arguments{
\item{downloadurl}{}
}
\description{
JSTOR Journal Master List
}
