% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master2wok.f.R
\name{master2wok.f}
\alias{master2wok.f}
\title{Scrape master list of journals indexed by Thompson Reuters Web of Knowledge. Subsample based on regular expression.}
\usage{

  master2wok.f(masterurl = "http://ip-science.thomsonreuters.com/cgi-bin/jrnlst/jlresults.cgi?PC=MASTER&mode=print",
  out = stop("Specify output directory for your project."), njour = NULL,
  npp = 500, check.for.saved.output = F, save = T, detect.usa = T,
  grep = NULL)
}
\arguments{
\item{grep}{}
}
\description{
Scrape master list of journals indexed by Thompson Reuters Web of Knowledge. Subsample based on regular expression.
}
