% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlc2mlk.f.R
\name{mlc2mlk.f}
\alias{mlc2mlk.f}
\title{Multilevel corpus 2 Multilevel K topic distribution}
\usage{
mlc2mlk.f(mlc, lev = c("doc", "par", "sen"))
}
\arguments{
\item{lev}{}
}
\description{
Multilevel corpus 2 Multilevel K topic distribution
}
