% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stmbow2lda.f.R
\name{stmbow2lda.f}
\alias{stmbow2lda.f}
\title{Takes stm style bag of words as input and gives LDA.}
\usage{
stmbow2lda.f(stmbow, out.dir, k = 0, alpha = NULL, sig.pri = 0.5,
  it = "Spectral", visualize.results = F, verbose = F, save.to.disk = F,
  check.for.saved.output = F, ...)
}
\arguments{
\item{stmbow}{stm formatted bag of words, as given by textProcessor command or stmbow importer.}

\item{k}{Number of topics.}

\item{alpha}{The alpha parameter must be greater than 0. Alpha < 1 assumes that each document is constructed from relatively few topics. Alpha > 1 assumes that each is constructed from many topics. If you aren't sure choose the convention: 50/k, which will also be the default if you specify nothing.}

\item{visualize.results}{Logical, if TRUE provides visualization tools to help interpret LDA output.}

\item{verbose}{Print iteration progress (very verbose!).}

\item{save.to.disk}{Save an .RData file of the object; helpful for lengthy estimations. May be buggy across environments.}

\item{...}{Other arguments to stm.}
}
\description{
Takes stm style bag of words as input and gives LDA.
}
