% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strdist.dend.picker.R
\name{strdist.dend.picker}
\alias{strdist.dend.picker}
\title{Interactive group picker for short strings clustered by string distance}
\usage{
strdist.dend.picker(hd, s1 = c(0, 0.075), s2 = c(0, 1),
  lightest.val = c(`>` = 0.5, `=` = 0.8), maxlist = 50,
  must.be.this.short.to.ride = 0.075, out = NULL, instruct = F)
}
\arguments{
\item{instruct}{}
}
\description{
Interactive group picker for short strings clustered by string distance
}
