% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx2pre.f.R
\name{ftx2pre.f}
\alias{ftx2pre.f}
\title{Full text to pre-processed text.}
\usage{
ftx2pre.f(ftx, pick = F, pl = "d/q/pl.RData", lvl = "sen", frq = 2,
  nch = 2)
}
\arguments{
\item{frq}{}
}
\description{
Full text to pre-processed text.
}
