% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master2jstorm.f.R
\name{master2jstorm.f}
\alias{master2jstorm.f}
\title{JSTOR Journal Master List}
\usage{
master2jstorm.f(pageurl,
  xp = "//*[@id=\\"content\\"]/div[1]/div/div[4]/div/table")
}
\arguments{
\item{xp}{}
}
\description{
JSTOR Journal Master List
}
